+++ src/types/brewery-brewery.ts
@@ -1,7 +1,9 @@
 // Auto-generated from src/modules/breweries/dto/brewery.dto.ts
 export type BreweryType = {
     name: string;
-    description?: string | undefined;
-    location?: string | undefined;
-    website?: string | undefined;
+    description?: (string | null) | undefined;
+    shortDescription?: (string | null) | undefined;
+    location?: (string | null) | undefined;
+    website?: (string | null) | undefined;
+    logo?: (string | null) | undefined;
 };