+++ src/types/brewery-updatebrewery.ts
@@ -1,7 +1,9 @@
 // Auto-generated from src/modules/breweries/dto/brewery.dto.ts
 export type UpdateBreweryType = {
     name?: string | undefined;
-    description?: (string | undefined) | undefined;
-    location?: (string | undefined) | undefined;
-    website?: (string | undefined) | undefined;
+    description?: ((string | null) | undefined) | undefined;
+    shortDescription?: ((string | null) | undefined) | undefined;
+    location?: ((string | null) | undefined) | undefined;
+    website?: ((string | null) | undefined) | undefined;
+    logo?: ((string | null) | undefined) | undefined;
 };