+++ src/types/category-category.ts
@@ -1,5 +1,5 @@
 // Auto-generated from src/modules/categories/dto/category.dto.ts
 export type CategoryType = {
     name: string;
-    description?: string | undefined;
+    description?: (string | null) | undefined;
 };