+++ src/types/category-categoryresponse.ts
@@ -1,8 +1,8 @@
 // Auto-generated from src/modules/categories/dto/category.dto.ts
 export type CategoryResponseType = {
-    name: string;
-    description?: string | undefined;
     id: string;
+    name: string;
+    description: string | null;
     createdAt: Date | string;
     updatedAt: Date | string;
 };