+++ src/types/category-updatecategory.ts
@@ -1,5 +1,5 @@
 // Auto-generated from src/modules/categories/dto/category.dto.ts
 export type UpdateCategoryType = {
     name?: string | undefined;
-    description?: (string | undefined) | undefined;
+    description?: ((string | null) | undefined) | undefined;
 };